pyproject.toml
[project]
name = "embedding-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.20",
    "flet>=0.28.3",
    "flet-desktop>=0.28.3",
    "langchain>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.3.7",
    "langgraph>=0.6.6",
    "markitdown[all]>=0.1.3",
    "ollama>=0.5.3",
    "paramiko>=4.0.0",
]
